requirements.txt
pyyaml
pytest
requests-mock
